build(board): modularize routing and API documentation
- Move route-related code into a new `route` module
- Reorganize imports and remove unused dependencies in `lib.rs`
- Extract API prefix constant and scalar setup into the route module
- Remove redundant code and improve server startup logic
- Update main function to properly call the board runner

This change improves code organization and maintainability by separating routing concerns
into their own module, making the core logic cleaner and more focused.
