+++ crates/board/src/lib.rs
@@ -1,18 +1,16 @@
 mod cmd;
 mod init;
+mod route;
 
 use std::net::SocketAddr;
 
-use axum::{Json, routing::get};
-use axum_responses::http::HttpResponse;
 use clap::Parser;
-use configrs::Config;
 use shadow_rs::shadow;
-use utoipa::OpenApi;
 use utoipa_axum::{router::OpenApiRouter, routes};
-use utoipa_scalar::{Scalar, Servable};
 
 use crate::init::{cfg_service, const_service, log_service};
+use configrs::Config;
+use route::*;
 
 shadow!(build);
 
@@ -32,8 +30,6 @@ pub async fn run() -> anyhow::Result<()> {
     Ok(())
 }
 
-const API_PREFIX: &str = "/api/v1";
-
 async fn start_server(cfg: &Config) -> anyhow::Result<()> {
     println!("Starting server with args: {cfg:#?}");
     log_service::init(cfg);
@@ -42,12 +38,11 @@ async fn start_server(cfg: &Config) -> anyhow::Result<()> {
     let listener = tokio::net::TcpListener::bind(addr).await?;
 
     let (router, api) = OpenApiRouter::new()
-        .routes(routes!(openapi))
+        .routes(routes!(route::openapi))
         .split_for_parts();
 
-    // let (router, api) = OpenApiRouter::with_openapi(ApiDoc::openapi()).nest(API_PREFIX, router);
-    let scalar_path = format!("{}/scalar", API_PREFIX);
-    let app = router.merge(Scalar::with_url(scalar_path.clone(), api));
+    let (scalar, scalar_path) = route::scalar(api);
+    let app = router.merge(scalar);
 
     println!("Server listening on http://{addr}");
     println!("Scalar listening on http://{addr}{}", &scalar_path);
@@ -55,36 +50,4 @@ async fn start_server(cfg: &Config) -> anyhow::Result<()> {
     axum::serve(listener, app.into_make_service())
         .await
         .map_err(anyhow::Error::new)
-}
-
-#[derive(OpenApi)]
-#[openapi(
-    info(
-        title = "API Docs",
-        description = "API documentation for the server",
-        version = build::PKG_VERSION
-    ),
-    tags((name=API_PREFIX,description="API v1")),
-    paths(openapi)
-)]
-struct ApiDoc;
-
-/// Return JSON version of an OpenAPI schema
-#[utoipa::path(
-    get,
-    path = format!("{}/openapi.json",API_PREFIX),
-    responses(
-        (status = 200, description = "JSON file", body = ())
-    )
-)]
-async fn openapi() -> Json<utoipa::openapi::OpenApi> {
-    Json(ApiDoc::openapi())
-}
-
-fn health() -> HttpResponse {
-    HttpResponse::Ok()
-}
-
-fn status() -> HttpResponse {
-    HttpResponse::Ok()
 }
