+++ crates/board/src/main.rs
@@ -1,4 +1,4 @@
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    Ok(())
+    board::run().await
 }
